Update to reflect changes to libgit2
These changes allow it to work with the latest libgit2 (9/4/2013)

+++ src/qgitindex.h
@@ -112,7 +112,7 @@ namespace LibQGit2
              * @param path path to search
              * @return an index >= 0 if found, -1 otherwise
              */
-            int find(const QString& path);
+            size_t find(const QString& path);
 
             /**
              * Add or update an index entry from a file in disk.
